+++ lib/_core.ts
