+++ lib/get_required_env.ts
