+++ lib/get_required_env_test.ts
