+++ lib/get_session_id_test.ts
@@ -1,6 +1,6 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 import { assertEquals } from "../dev_deps.ts";
-import { SITE_COOKIE_NAME } from "./core.ts";
+import { SITE_COOKIE_NAME } from "./_core.ts";
 import { getSessionId } from "./get_session_id.ts";
 
 Deno.test("getSessionId() returns undefined when cookie is not defined", () => {
