+++ mod.ts
@@ -1,20 +1,20 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
-export * from "./src/clear_oauth_sessions_and_tokens.ts";
-export * from "./src/get_session_access_token.ts";
-export * from "./src/handle_callback.ts";
-export * from "./src/get_session_id.ts";
-export * from "./src/sign_in.ts";
-export * from "./src/sign_out.ts";
-export * from "./src/create_auth0_oauth_config.ts";
-export * from "./src/create_discord_oauth_config.ts";
-export * from "./src/create_dropbox_oauth_config.ts";
-export * from "./src/create_facebook_oauth_config.ts";
-export * from "./src/create_github_oauth_config.ts";
-export * from "./src/create_gitlab_oauth_config.ts";
-export * from "./src/create_google_oauth_config.ts";
-export * from "./src/create_notion_oauth_config.ts";
-export * from "./src/create_okta_oauth_config.ts";
-export * from "./src/create_patreon_oauth_config.ts";
-export * from "./src/create_slack_oauth_config.ts";
-export * from "./src/create_spotify_oauth_config.ts";
-export * from "./src/create_twitter_oauth_config.ts";
+export * from "./lib/clear_oauth_sessions_and_tokens.ts";
+export * from "./lib/get_session_access_token.ts";
+export * from "./lib/handle_callback.ts";
+export * from "./lib/get_session_id.ts";
+export * from "./lib/sign_in.ts";
+export * from "./lib/sign_out.ts";
+export * from "./lib/create_auth0_oauth_config.ts";
+export * from "./lib/create_discord_oauth_config.ts";
+export * from "./lib/create_dropbox_oauth_config.ts";
+export * from "./lib/create_facebook_oauth_config.ts";
+export * from "./lib/create_github_oauth_config.ts";
+export * from "./lib/create_gitlab_oauth_config.ts";
+export * from "./lib/create_google_oauth_config.ts";
+export * from "./lib/create_notion_oauth_config.ts";
+export * from "./lib/create_okta_oauth_config.ts";
+export * from "./lib/create_patreon_oauth_config.ts";
+export * from "./lib/create_slack_oauth_config.ts";
+export * from "./lib/create_spotify_oauth_config.ts";
+export * from "./lib/create_twitter_oauth_config.ts";
